+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Delay zero value bug fixed
+- Jitter zero value bug fixed
 ## [1.1] - 2026-06-03
 ### Added
 - `RUBBER_DUCK` mode