+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - Delay zero value bug fixed
 - Jitter zero value bug fixed
+- Test system modified
 ## [1.1] - 2026-06-03
 ### Added
 - `RUBBER_DUCK` mode