+++ CHANGELOG.md
@@ -5,8 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `TypioContext.random` property
 ### Changed
+- Global random state pollution bug fixed
 - Test system modified
+- `README.md` updated
 - `--mode` CLI argument modified
 ## [1.2] - 2026-08-17
 ### Added