+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `README.md` updated
 - `--mode` CLI argument modified
 - Validation system modified
+- `functions.py` renamed to `core.py`
 ## [1.2] - 2026-08-17
 ### Added
 - `text` CLI argument